Update README.md
Set-up: For this assignment, grab the starter code. To complete these problems, edit hw5.rkt to replace all occurrences of “CHANGE” with your solutions.

Do not use any mutation (set!, set-mcar!, etc.) anywhere in the assignment.

Overview: This homework has to do with mupl (a Made Up Programming Language). mupl programs are written directly in Racket by using the constructors de ned by the structs de ned at the beginning of hw5.rkt. This is the de nition of mupl’s syntax:

If s is a Racket string, then (var s) is a mupl expression (a variable use).

If n is a Racket integer, then (int n) is a mupl expression (a constant).

If e1 and e2 are mupl expressions, then (add e1 e2) is a mupl expression (an addition).